FER-VIT-Code/requirements.txt
torch==2.2.2
torchvision==0.17.2
transformers==4.40.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
scikit-learn==1.5.0
opencv-python==4.9.0.80
Pillow==10.3.0
tqdm==4.66.4
scipy==1.13.0
piq==0.8.0
lpips==0.1.4
